Fix economy failure for positive amounts, overhaul EconomyManager definition and JavaDocs Core interfaces now use BigDecimal internally to eliminate possible rounding issues.

+++ mywarp-bukkit/src/main/java/me/taylorkelly/mywarp/bukkit/BukkitSettings.java
@@ -38,6 +38,7 @@ import org.slf4j.Logger;
 
 import java.io.File;
 import java.io.IOException;
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Locale;
@@ -348,14 +349,25 @@ public class BukkitSettings implements Settings {
    * @return a FeeBundle out of the given values
    */
   private FeeBundle toFeeBundle(String identifier, ConfigurationSection values) {
-    return new FeeBundle(identifier, values.getDouble("assets", 0), values.getDouble("create", 0),
-                         values.getDouble("createPrivate", 0), values.getDouble("delete", 0),
-                         values.getDouble("give", 0), values.getDouble("help", 0), values.getDouble("info", 0),
-                         values.getDouble("invite", 0), values.getDouble("list", 0), values.getDouble("point", 0),
-                         values.getDouble("private", 0), values.getDouble("public", 0), values.getDouble("uninvite", 0),
-                         values.getDouble("update", 0), values.getDouble("warpPlayer", 0),
-                         values.getDouble("warpSignCreate", 0), values.getDouble("warpSignUse", 0),
-                         values.getDouble("warpTo", 0), values.getDouble("welcome", 0));
+    return new FeeBundle(identifier, BigDecimal.valueOf(values.getDouble("assets", 0)),
+                         BigDecimal.valueOf(values.getDouble("create", 0)),
+                         BigDecimal.valueOf(values.getDouble("createPrivate", 0)),
+                         BigDecimal.valueOf(values.getDouble("delete", 0)),
+                         BigDecimal.valueOf(values.getDouble("give", 0)),
+                         BigDecimal.valueOf(values.getDouble("help", 0)),
+                         BigDecimal.valueOf(values.getDouble("info", 0)),
+                         BigDecimal.valueOf(values.getDouble("invite", 0)),
+                         BigDecimal.valueOf(values.getDouble("list", 0)),
+                         BigDecimal.valueOf(values.getDouble("point", 0)),
+                         BigDecimal.valueOf(values.getDouble("private", 0)),
+                         BigDecimal.valueOf(values.getDouble("public", 0)),
+                         BigDecimal.valueOf(values.getDouble("uninvite", 0)),
+                         BigDecimal.valueOf(values.getDouble("update", 0)),
+                         BigDecimal.valueOf(values.getDouble("warpPlayer", 0)),
+                         BigDecimal.valueOf(values.getDouble("warpSignCreate", 0)),
+                         BigDecimal.valueOf(values.getDouble("warpSignUse", 0)),
+                         BigDecimal.valueOf(values.getDouble("warpTo", 0)),
+                         BigDecimal.valueOf(values.getDouble("welcome", 0)));
   }
 
   /**
